+++ src/main/java/com/planbase/pdf/layoutmanager/CellStyle.java
@@ -33,16 +33,16 @@ public class CellStyle {
                 if (outer.lte(inner)) { return null; }
                 return Padding.of(0, outer.x() - inner.x(), outer.y() - inner.y(), 0);
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) { return 0f; }
+            @Override public double leftOffset(double outerWidth, double innerWidth) { return 0; }
         },
         TOP_CENTER {
             @Override public Padding calcPadding(XyDim outer, XyDim inner) {
                 if (outer.lte(inner)) { return null; }
-                float dx = (outer.x() - inner.x()) / 2;
+                double dx = (outer.x() - inner.x()) / 2;
                 return Padding.of(0, dx, outer.y() - inner.y(), dx);
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) {
-                return (innerWidth >= outerWidth) ? 0f : (outerWidth - innerWidth) / 2;
+            @Override public double leftOffset(double outerWidth, double innerWidth) {
+                return (innerWidth >= outerWidth) ? 0 : (outerWidth - innerWidth) / 2;
             }
         },
         TOP_RIGHT {
@@ -50,37 +50,37 @@ public class CellStyle {
                 if (outer.lte(inner)) { return null; }
                 return Padding.of(0, 0, outer.y() - inner.y(), outer.x() - inner.x());
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) {
-                return (innerWidth >= outerWidth) ? 0f : (outerWidth - innerWidth);
+            @Override public double leftOffset(double outerWidth, double innerWidth) {
+                return (innerWidth >= outerWidth) ? 0 : (outerWidth - innerWidth);
             }
         },
         MIDDLE_LEFT {
             @Override public Padding calcPadding(XyDim outer, XyDim inner) {
                 if (outer.lte(inner)) { return null; }
-                float dy = (outer.y() - inner.y()) / 2;
+                double dy = (outer.y() - inner.y()) / 2;
                 return Padding.of(dy, outer.x() - inner.x(), dy, 0);
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) { return 0f; }
+            @Override public double leftOffset(double outerWidth, double innerWidth) { return 0; }
         },
         MIDDLE_CENTER {
             @Override public Padding calcPadding(XyDim outer, XyDim inner) {
                 if (outer.lte(inner)) { return null; }
-                float dx = (outer.x() - inner.x()) / 2;
-                float dy = (outer.y() - inner.y()) / 2;
+                double dx = (outer.x() - inner.x()) / 2;
+                double dy = (outer.y() - inner.y()) / 2;
                 return Padding.of(dy, dx, dy, dx);
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) {
-                return (innerWidth >= outerWidth) ? 0f : (outerWidth - innerWidth) / 2;
+            @Override public double leftOffset(double outerWidth, double innerWidth) {
+                return (innerWidth >= outerWidth) ? 0 : (outerWidth - innerWidth) / 2;
             }
         },
         MIDDLE_RIGHT {
             @Override public Padding calcPadding(XyDim outer, XyDim inner) {
                 if (outer.lte(inner)) { return null; }
-                float dy = (outer.y() - inner.y()) / 2;
+                double dy = (outer.y() - inner.y()) / 2;
                 return Padding.of(dy, 0, dy, outer.x() - inner.x());
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) {
-                return (innerWidth >= outerWidth) ? 0f : (outerWidth - innerWidth);
+            @Override public double leftOffset(double outerWidth, double innerWidth) {
+                return (innerWidth >= outerWidth) ? 0 : (outerWidth - innerWidth);
             }
         },
         BOTTOM_LEFT {
@@ -88,20 +88,20 @@ public class CellStyle {
                 if (outer.lte(inner)) { return null; }
                 return Padding.of(outer.y() - inner.y(), outer.x() - inner.x(), 0, 0);
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) { return 0f; }
+            @Override public double leftOffset(double outerWidth, double innerWidth) { return 0; }
         },
         BOTTOM_CENTER {
             @Override public Padding calcPadding(XyDim outer, XyDim inner) {
 //                System.out.println("\t\t\tcalcPadding(o=" + outer + " i=" + inner + ")");
                 if (outer.lte(inner)) { return null; }
-                float dx = (outer.x() - inner.x()) / 2;
+                double dx = (outer.x() - inner.x()) / 2;
 //                System.out.println("\t\t\tcalcPadding() dx=" + dx);
                 // Like HTML it's top, right, bottom, left
 //                System.out.println("\t\t\tcalcPadding() outer.y() - inner.y()=" + (outer.y() - inner.y()));
                 return Padding.of(outer.y() - inner.y(), dx, 0, dx);
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) {
-                return (innerWidth >= outerWidth) ? 0f : (outerWidth - innerWidth) / 2;
+            @Override public double leftOffset(double outerWidth, double innerWidth) {
+                return (innerWidth >= outerWidth) ? 0 : (outerWidth - innerWidth) / 2;
             }
         },
         BOTTOM_RIGHT {
@@ -109,8 +109,8 @@ public class CellStyle {
                 if (outer.lte(inner)) { return null; }
                 return Padding.of(outer.y() - inner.y(), 0, 0, outer.x() - inner.x());
             }
-            @Override public float leftOffset(float outerWidth, float innerWidth) {
-                return (innerWidth >= outerWidth) ? 0f : (outerWidth - innerWidth);
+            @Override public double leftOffset(double outerWidth, double innerWidth) {
+                return (innerWidth >= outerWidth) ? 0 : (outerWidth - innerWidth);
             }
         };
 
@@ -119,7 +119,7 @@ public class CellStyle {
         calculates additional padding to apply.
         */
         public abstract Padding calcPadding(XyDim outer, XyDim inner);
-        public abstract float leftOffset(float outerWidth, float innerWidth);
+        public abstract double leftOffset(double outerWidth, double innerWidth);
     }
 
     public static final Align DEFAULT_ALIGN = Align.TOP_LEFT;
@@ -146,7 +146,7 @@ public class CellStyle {
     public BorderStyle borderStyle() { return borderStyle; }
 
     // NOTE: This lincluded the padding!
-//    public float calcLeftX(float x, float spareRoom) {
+//    public double calcLeftX(double x, double spareRoom) {
 //        return HorizAlign.LEFT == align ? x + padding.left() :
 //               HorizAlign.CENTER == align ? x + (spareRoom / 2) :
 //               HorizAlign.RIGHT == align ? x + spareRoom - padding.right() :
@@ -159,7 +159,7 @@ public class CellStyle {
     public CellStyle borderStyle(BorderStyle bs) {
         return new Builder(this).borderStyle(bs).build();
     }
-    public Cell.Builder cellBuilder(float w) {
+    public Cell.Builder cellBuilder(double w) {
         return Cell.builder(this, w);
     }
